+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/domain/port/output/repository/AdditionalSkillCategoryRepository.java
@@ -1,7 +1,0 @@
-package fr.avenirsesr.portfolio.additionalskill.domain.port.output.repository;
-
-import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkillCategory;
-import fr.avenirsesr.portfolio.common.data.domain.port.output.repository.GenericRepositoryPort;
-
-public interface AdditionalSkillCategoryRepository
-    extends GenericRepositoryPort<AdditionalSkillCategory> {}