+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/infrastructure/adapter/repository/AdditionalSkillDatabaseRepository.java
@@ -1,13 +1,12 @@
 package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.repository;
 
 import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkill;
-import fr.avenirsesr.portfolio.additionalskill.domain.model.enums.EAdditionalSkillType;
 import fr.avenirsesr.portfolio.additionalskill.domain.port.output.repository.AdditionalSkillRepository;
 import fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.mapper.AdditionalSkillMapper;
 import fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model.AdditionalSkillEntity;
-import fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.specification.AdditionalSkillSpecification;
 import fr.avenirsesr.portfolio.common.data.infrastructure.adapter.repository.GenericJpaRepositoryAdapter;
-import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -26,14 +25,9 @@ public class AdditionalSkillDatabaseRepository
   }
 
   @Override
-  public List<AdditionalSkill> findAllByExternalId(List<String> skillCodes) {
-    return jpaRepository.findAll(AdditionalSkillSpecification.hasExternalId(skillCodes)).stream()
-        .map(AdditionalSkillMapper::toDomain)
-        .toList();
-  }
-
-  @Override
-  public int countAll(EAdditionalSkillType type) {
-    return jpaRepository.findAll(AdditionalSkillSpecification.hasType(type)).size();
+  public Optional<AdditionalSkill> findByExternalSkillId(UUID externalSkillId) {
+    return jpaRepository
+        .findByExternalSkillId(externalSkillId)
+        .map(AdditionalSkillMapper::toDomain);
   }
 }