+++ src/main/java/fr/avenirsesr/portfolio/interoperability/additionalskill/casoc/domain/port/input/CasocService.java
@@ -1,8 +1,0 @@
-package fr.avenirsesr.portfolio.interoperability.additionalskill.casoc.domain.port.input;
-
-import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkill;
-import java.util.List;
-
-public interface CasocService {
-  List<AdditionalSkill> syncSkills();
-}