+++ src/main/java/fr/avenirsesr/portfolio/interoperability/additionalskill/casol/domain/service/CompetenceReader.java
@@ -1,8 +1,0 @@
-package fr.avenirsesr.portfolio.interoperability.additionalskill.casol.domain.service;
-
-import fr.avenirsesr.portfolio.interoperability.additionalskill.casol.domain.model.Competence;
-import java.util.List;
-
-public interface CompetenceReader {
-  List<Competence> readCompetences();
-}