+++ src/main/java/fr/avenirsesr/portfolio/interoperability/additionalskill/xxi/domain/model/Competence.java
@@ -1,3 +1,0 @@
-package fr.avenirsesr.portfolio.interoperability.additionalskill.xxi.domain.model;
-
-public record Competence(int id, String libelle, Category category) {}