+++ src/main/java/fr/avenirsesr/portfolio/student/progress/domain/data/AdditionalSkillProgressDetails.java
@@ -1,9 +1,11 @@
 package fr.avenirsesr.portfolio.student.progress.domain.data;
 
+import fr.avenirsesr.portfolio.common.externalskill.application.adapter.dto.ExternalSkillCategoryDTO;
 import fr.avenirsesr.portfolio.student.progress.domain.model.AdditionalSkillProgress;
 import fr.avenirsesr.portfolio.trace.domain.data.TraceWithProjectNameData;
 import java.util.List;
 
 public record AdditionalSkillProgressDetails(
     AdditionalSkillProgress additionalSkillProgress,
-    List<TraceWithProjectNameData> tracesWithProjectName) {}
+    List<TraceWithProjectNameData> tracesWithProjectName,
+    List<ExternalSkillCategoryDTO> externalCategories) {}