+++ src/main/java/fr/avenirsesr/portfolio/user/domain/port/input/UserService.java
@@ -1,13 +1,12 @@
 package fr.avenirsesr.portfolio.user.domain.port.input;
 
-import fr.avenirsesr.portfolio.common.data.domain.model.User;
 import fr.avenirsesr.portfolio.common.data.domain.model.enums.EUserCategory;
+import fr.avenirsesr.portfolio.common.user.domain.port.output.BaseUserService;
 import fr.avenirsesr.portfolio.user.domain.data.UserPhotosData;
 import fr.avenirsesr.portfolio.user.domain.data.UserProfileOverviewData;
 import java.util.UUID;
 
-public interface UserService {
-  User getUser(UUID id);
+public interface UserService extends BaseUserService {
 
   UserPhotosData getUserPhotos(UUID userId, EUserCategory userCategory);